removed worker specific snapshot/bundle
