--- js/workers_guest.ts
@@ -1,70 +1,0 @@
-// Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import { sendAsync, sendSync } from "./dispatch";
-import * as msg from "gen/cli/msg_generated";
-import * as flatbuffers from "./flatbuffers";
-import { assert, log } from "./util";
-import { encodeMessage, decodeMessage } from "./workers";
-import { window } from "./window";
-
-export let onmessage: (e: { data: any }) => void = (): void => {};
-
-export function postMessage(data: any): void {
-  const dataIntArray = encodeMessage(data);
-  const builder = flatbuffers.createBuilder();
-  msg.WorkerPostMessage.startWorkerPostMessage(builder);
-  const inner = msg.WorkerPostMessage.endWorkerPostMessage(builder);
-  const baseRes = sendSync(
-    builder,
-    msg.Any.WorkerPostMessage,
-    inner,
-    dataIntArray
-  );
-  assert(baseRes != null);
-}
-
-export async function getMessage(): Promise<any> {
-  log("getMessage");
-  const builder = flatbuffers.createBuilder();
-  msg.WorkerGetMessage.startWorkerGetMessage(builder);
-  const inner = msg.WorkerGetMessage.endWorkerGetMessage(builder);
-  const baseRes = await sendAsync(builder, msg.Any.WorkerGetMessage, inner);
-  assert(baseRes != null);
-  assert(
-    msg.Any.WorkerGetMessageRes === baseRes!.innerType(),
-    `base.innerType() unexpectedly is ${baseRes!.innerType()}`
-  );
-  const res = new msg.WorkerGetMessageRes();
-  assert(baseRes!.inner(res) != null);
-
-  const dataArray = res.dataArray();
-  if (dataArray != null) {
-    return decodeMessage(dataArray);
-  } else {
-    return null;
-  }
-}
-
-export let isClosing = false;
-
-export function workerClose(): void {
-  isClosing = true;
-}
-
-export async function workerMain(): Promise<void> {
-  log("workerMain");
-
-  while (!isClosing) {
-    const data = await getMessage();
-    if (data == null) {
-      log("workerMain got null message. quitting.");
-      break;
-    }
-    if (window["onmessage"]) {
-      const event = { data };
-      window.onmessage(event);
-    } else {
-      break;
-    }
-  }
-}
