+++ tsconfig.json
@@ -27,7 +27,6 @@
     "core/core.d.ts",
     "core/libdeno/libdeno.d.ts",
     "js/main.ts",
-    "js/workers_init.ts",
     "js/compiler.ts"
   ]
 }
